+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ v0.11.1 (in development)
 ------------------------
 Features:
 
+* Allow adding a source with a base URI of ``None`` to match full URIs as the ``relative_path``
 * ``JSONPointer`` and ``RelativeJSONPointer`` now have class attributes defining
   the exceptions that they use, which can be overidden in subclasses
 * Cached properties for accessing document and resource root schemas from subschemas